improve documentation
